+++ src/commands/info.ts
@@ -22,7 +22,7 @@ import nuxiPkg from '../../package.json'
 export default defineCommand({
   meta: {
     name: 'info',
-    description: 'Get information about nuxt project',
+    description: 'Get information about Nuxt project',
   },
   args: {
     ...sharedArgs,
@@ -32,7 +32,7 @@ export default defineCommand({
     // Resolve rootDir
     const cwd = resolve(ctx.args.cwd || ctx.args.rootDir || '.')
 
-    // Load nuxt.config
+    // Load Nuxt config
     const nuxtConfig = getNuxtConfig(cwd)
 
     // Find nearest package.json
@@ -53,9 +53,10 @@ export default defineCommand({
         })
         .join(', ')
 
-    // Check nuxt version
+    // Check Nuxt version
     const nuxtVersion =
       getDepVersion('nuxt') ||
+      getDepVersion('nuxt-nightly') ||
       getDepVersion('nuxt-edge') ||
       getDepVersion('nuxt3') ||
       '-'
@@ -63,10 +64,10 @@ export default defineCommand({
     const builder = !isLegacy
       ? nuxtConfig.builder /* latest schema */ || '-'
       : nuxtConfig.bridge?.vite
-      ? 'vite' /* bridge vite implementation */
-      : nuxtConfig.buildModules?.includes('nuxt-vite')
-      ? 'vite' /* nuxt-vite */
-      : 'webpack'
+        ? 'vite' /* bridge vite implementation */
+        : nuxtConfig.buildModules?.includes('nuxt-vite')
+          ? 'vite' /* nuxt-vite */
+          : 'webpack'
 
     let packageManager: keyof typeof packageManagerLocks | 'unknown' | null =
       getPackageManager(cwd)