+++ src/commands/module/add.ts
@@ -5,11 +5,21 @@ import { existsSync } from 'node:fs'
 import { loadFile, writeFile, parseModule, ProxifiedModule } from 'magicast'
 import consola from 'consola'
 import { addDependency } from 'nypm'
+import { $fetch } from 'ofetch'
+import {
+  NuxtModule,
+  checkNuxtCompatibility,
+  fetchModules,
+  getNuxtVersion,
+  getProjectPackage,
+} from './_utils'
+import { satisfies } from 'semver'
+import { colors } from 'consola/utils'
 
 export default defineCommand({
   meta: {
-    name: 'module',
-    description: 'Manage Nuxt Modules',
+    name: 'add',
+    description: 'Add Nuxt modules',
   },
   args: {
     ...sharedArgs,
@@ -28,19 +38,52 @@ export default defineCommand({
   },
   async setup(ctx) {
     const cwd = resolve(ctx.args.cwd || '.')
+    const projectPkg = await getProjectPackage(cwd)
 
-    // TODO: Resolve and validate npm package name first
-    const npmPackage = ctx.args.moduleName
+    if (!projectPkg.dependencies?.nuxt && !projectPkg.devDependencies?.nuxt) {
+      consola.warn(`No \`nuxt\` dependency detected in \`${cwd}\`.`)
+      const shouldContinue = await consola.prompt(
+        `Do you want to continue anyway?`,
+        {
+          type: 'confirm',
+          initial: false,
+        },
+      )
+      if (shouldContinue !== true) {
+        return false
+      }
+    }
+
+    const r = await resolveModule(ctx.args.moduleName, cwd)
+    if (r === false) {
+      return
+    }
 
     // Add npm dependency
     if (!ctx.args.skipInstall) {
-      consola.info(`Installing dev dependency \`${npmPackage}\``)
-      await addDependency(npmPackage, { cwd, dev: true }).catch((err) => {
-        consola.error(err)
-        consola.error(
-          `Please manually install \`${npmPackage}\` as a dev dependency`,
-        )
-      })
+      const isDev = Boolean(projectPkg.devDependencies?.nuxt)
+      consola.info(
+        `Installing \`${r.pkg}\`${isDev ? ' development' : ''} dependency`,
+      )
+      const res = await addDependency(r.pkg, { cwd, dev: isDev }).catch(
+        (error) => {
+          consola.error(error)
+          return consola.prompt(
+            `Install failed for ${colors.cyan(
+              r.pkg,
+            )}. Do you want to continue adding the module to ${colors.cyan(
+              'nuxt.config',
+            )}?`,
+            {
+              type: 'confirm',
+              initial: false,
+            },
+          )
+        },
+      )
+      if (res === false) {
+        return
+      }
     }
 
     // Update nuxt.config.ts
@@ -49,18 +92,17 @@ export default defineCommand({
         if (!config.modules) {
           config.modules = []
         }
-        for (let i = 0; i < config.modules.length; i++) {
-          if (config.modules[i] === npmPackage) {
-            consola.info(`\`${npmPackage}\` is already in the \`modules\``)
-            return
-          }
+
+        if (config.modules.includes(r.pkgName)) {
+          consola.info(`\`${r.pkgName}\` is already in the \`modules\``)
+          return
         }
-        consola.info(`Adding \`${npmPackage}\` to the \`modules\``)
-        config.modules.push(npmPackage)
+        consola.info(`Adding \`${r.pkgName}\` to the \`modules\``)
+        config.modules.push(r.pkgName)
       }).catch((err) => {
         consola.error(err)
         consola.error(
-          `Please manually add \`${npmPackage}\` to the \`modules\` in \`nuxt.config.ts\``,
+          `Please manually add \`${r.pkgName}\` to the \`modules\` in \`nuxt.config.ts\``,
         )
       })
     }
@@ -101,4 +143,124 @@ function getDefaultNuxtConfig() {
 export default defineNuxtConfig({
   modules: []
 })`
+}
+
+// Based on https://github.com/dword-design/package-name-regex
+const packageRegex =
+  /^(@[a-z0-9-~][a-z0-9-._~]*\/)?([a-z0-9-~][a-z0-9-._~]*)(@[^@]+)?$/
+
+async function resolveModule(
+  moduleName: string,
+  cwd: string,
+): Promise<
+  | false
+  | {
+      nuxtModule?: NuxtModule
+      pkg: string
+      pkgName: string
+      pkgVersion: string
+    }
+> {
+  let pkgName = moduleName
+  let pkgVersion: string | undefined
+
+  const reMatch = moduleName.match(packageRegex)
+  if (reMatch) {
+    if (reMatch[3]) {
+      pkgName = `${reMatch[1] || ''}${reMatch[2] || ''}`
+      pkgVersion = reMatch[3].slice(1)
+    }
+  } else {
+    consola.error(`Invalid package name \`${pkgName}\`.`)
+    return false
+  }
+
+  const modulesDB = await fetchModules().catch((err) => {
+    consola.warn('Cannot search in the Nuxt Modules database: ' + err)
+    return []
+  })
+
+  const matchedModule = modulesDB.find(
+    (module) =>
+      module.name === moduleName ||
+      module.npm === pkgName ||
+      module.aliases?.includes(pkgName),
+  )
+
+  if (matchedModule?.npm) {
+    pkgName = matchedModule.npm
+  }
+
+  if (matchedModule && matchedModule.compatibility.nuxt) {
+    // Get local Nuxt version
+    const nuxtVersion = await getNuxtVersion(cwd)
+
+    // Check for Module Compatibility
+    if (!checkNuxtCompatibility(matchedModule, nuxtVersion)) {
+      consola.warn(
+        `The module \`${pkgName}\` is not compatible with Nuxt \`${nuxtVersion}\` (requires \`${matchedModule.compatibility.nuxt}\`)`,
+      )
+      const shouldContinue = await consola.prompt(
+        'Do you want to continue installing incompatible version?',
+        {
+          type: 'confirm',
+          initial: false,
+        },
+      )
+      if (shouldContinue !== true) {
+        return false
+      }
+    }
+
+    // Match corresponding version of module for local Nuxt version
+    const versionMap = matchedModule.compatibility.versionMap
+    if (versionMap) {
+      for (const [_nuxtVersion, _moduleVersion] of Object.entries(versionMap)) {
+        if (satisfies(nuxtVersion, _nuxtVersion)) {
+          if (!pkgVersion) {
+            pkgVersion = _moduleVersion
+          } else {
+            consola.warn(
+              `Recommended version of \`${pkgName}\` for Nuxt \`${nuxtVersion}\` is \`${_moduleVersion}\` but you have requested \`${pkgVersion}\``,
+            )
+            pkgVersion = await consola.prompt('Choose a version:', {
+              type: 'select',
+              options: [_moduleVersion, pkgVersion],
+            })
+          }
+          break
+        }
+      }
+    }
+  }
+
+  // Fetch package on npm
+  pkgVersion = pkgVersion || 'latest'
+  const pkg = await $fetch(
+    `https://registry.npmjs.org/${pkgName}/${pkgVersion}`,
+  )
+  const pkgDependencies = Object.assign(
+    pkg.dependencies || {},
+    pkg.devDependencies || {},
+  )
+  if (!pkgDependencies['nuxt'] && !pkgDependencies['nuxt-edge']) {
+    consola.warn(`It seems that \`${pkgName}\` is not a Nuxt module.`)
+    const shouldContinue = await consola.prompt(
+      `Do you want to continue installing \`${pkgName}\` anyway?`,
+      {
+        type: 'confirm',
+        initial: false,
+      },
+    )
+    if (shouldContinue !== true) {
+      return false
+    }
+  }
+
+  return {
+    nuxtModule: matchedModule,
+    pkg: `${pkgName}@${pkgVersion}`,
+    pkgName,
+    pkgVersion,
+  }
 }