+++ src/main.ts
@@ -1,4 +1,5 @@
 import { defineCommand } from 'citty'
+import { provider } from 'std-env'
 import { commands } from './commands'
 import { setupGlobalConsole } from './utils/console'
 import { checkEngines } from './utils/engines'
@@ -19,10 +20,13 @@ export const main = defineCommand({
     setupGlobalConsole({ dev })
 
     // Check Node.js version and CLI updates in background
-    const backgroundTasks = Promise.all([
-      checkEngines(),
-      // checkForUpdates(),
-    ]).catch((err) => console.error(err))
+    let backgroundTasks: Promise<any> | undefined
+    if (command !== '_dev' && provider !== 'stackblitz') {
+      backgroundTasks = Promise.all([
+        checkEngines(),
+        // checkForUpdates(),
+      ]).catch((err) => console.error(err))
+    }
 
     // Avoid background check to fix prompt issues
     if (command === 'init') {