+++ src/utils/banner.ts
@@ -1,21 +1,17 @@
-import clear from 'clear'
 import { bold, gray, green } from 'colorette'
-import nuxiPkg from '../../package.json'
 import { tryRequireModule } from './cjs'
-
-export function showBanner(_clear?: boolean) {
-  if (_clear) {
-    clear()
-  }
-  console.log(gray(`Nuxt CLI ${bold(nuxiPkg.version)}`))
-}
 
 export function showVersions(cwd: string) {
   const getPkgVersion = (pkg: string) => {
     return tryRequireModule(`${pkg}/package.json`, cwd)?.version || ''
   }
-  const nuxtVersion = getPkgVersion('nuxt') || getPkgVersion('nuxt-edge')
-  const nitroVersion = getPkgVersion('nitropack')
+  const nuxtVersion =
+    getPkgVersion('nuxt') ||
+    getPkgVersion('nuxt-nightly') ||
+    getPkgVersion('nuxt3') ||
+    getPkgVersion('nuxt-edge')
+  const nitroVersion =
+    getPkgVersion('nitropack') || getPkgVersion('nitropack-edge')
   console.log(
     gray(
       green(`Nuxt ${bold(nuxtVersion)}`) +