+++ src/utils/dev.ts
@@ -1,21 +1,36 @@
 import type { RequestListener, ServerResponse } from 'node:http'
 import EventEmitter from 'node:events'
-import { relative, resolve } from 'pathe'
+import { relative, resolve, join } from 'pathe'
 import chokidar from 'chokidar'
 import { consola } from 'consola'
 import { debounce } from 'perfect-debounce'
 import { toNodeListener } from 'h3'
-import { HTTPSOptions, listen, Listener } from 'listhen'
+import {
+  HTTPSOptions,
+  ListenURL,
+  listen,
+  Listener,
+  ListenOptions,
+} from 'listhen'
 import type { Nuxt, NuxtConfig } from '@nuxt/schema'
 import { loadKit } from '../utils/kit'
 import { loadNuxtManifest, writeNuxtManifest } from '../utils/nuxt'
 import { clearBuildDir } from '../utils/fs'
 import { importModule } from './esm'
+import { AddressInfo } from 'node:net'
 
 export type NuxtDevIPCMessage =
   | { type: 'nuxt:internal:dev:ready'; port: number }
   | { type: 'nuxt:internal:dev:loading'; message: string }
   | { type: 'nuxt:internal:dev:restart' }
+
+export interface NuxtDevContext {
+  proxy?: {
+    url?: string
+    urls?: ListenURL[]
+    https?: boolean | HTTPSOptions
+  }
+}
 
 export interface NuxtDevServerOptions {
   cwd: string
@@ -23,22 +38,48 @@ export interface NuxtDevServerOptions {
   dotenv: boolean
   clear: boolean
   overrides: NuxtConfig
-  https?: boolean | HTTPSOptions
   port?: string | number
   loadingTemplate?: ({ loading }: { loading: string }) => string
+  devContext: NuxtDevContext
 }
 
-export async function createNuxtDevServer(options: NuxtDevServerOptions) {
+export async function createNuxtDevServer(
+  options: NuxtDevServerOptions,
+  listenOptions?: Partial<ListenOptions>,
+) {
+  // Initialize dev server
   const devServer = new NuxtDevServer(options)
-  devServer.listener = await listen(devServer.handler, {
-    port: options.port ?? 0,
-    hostname: '127.0.0.1',
-    showURL: false,
-  })
+
+  // Attach internal listener
+  devServer.listener = await listen(
+    devServer.handler,
+    listenOptions || {
+      port: options.port ?? 0,
+      hostname: '127.0.0.1',
+      showURL: false,
+    },
+  )
+
+  // Merge interface with public context
+  // @ts-expect-error
+  devServer.listener._url = devServer.listener.url
+  if (options.devContext.proxy?.url) {
+    devServer.listener.url = options.devContext.proxy.url
+  }
+  if (options.devContext.proxy?.urls) {
+    const _getURLs = devServer.listener.getURLs.bind(devServer.listener)
+    devServer.listener.getURLs = async () =>
+      Array.from(
+        new Set([...options.devContext.proxy!.urls!, ...(await _getURLs())]),
+      )
+  }
+
   return devServer
 }
 
-const RESTART_RE = /^(nuxt\.config\.(js|ts|mjs|cjs)|\.nuxtignore|\.nuxtrc)$/
+// https://regex101.com/r/7HkR5c/1
+const RESTART_RE =
+  /^(nuxt\.config\.[a-z0-9]+|\.nuxtignore|\.nuxtrc|\.config\/nuxt(\.config)?\.[a-z0-9]+)$/
 
 class NuxtDevServer extends EventEmitter {
   private _handler?: RequestListener
@@ -107,14 +148,14 @@ class NuxtDevServer extends EventEmitter {
       consola.error(`Cannot ${reload ? 'restart' : 'start'} nuxt: `, error)
       this._handler = undefined
       this._loadingMessage =
-        'Error while loading nuxt. Please check console and fix errors.'
+        'Error while loading Nuxt. Please check console and fix errors.'
       this.emit('loading', this._loadingMessage)
     }
   }
 
   async _load(reload?: boolean, reason?: string) {
     const action = reload ? 'Restarting' : 'Starting'
-    this._loadingMessage = `${reason ? reason + '. ' : ''}${action} nuxt...`
+    this._loadingMessage = `${reason ? reason + '. ' : ''}${action} Nuxt...`
     this._handler = undefined
     this.emit('loading', this._loadingMessage)
     if (reload) {
@@ -130,7 +171,7 @@ class NuxtDevServer extends EventEmitter {
 
     const kit = await loadKit(this.options.cwd)
     this._currentNuxt = await kit.loadNuxt({
-      rootDir: this.options.cwd,
+      cwd: this.options.cwd,
       dev: true,
       ready: false,
       overrides: {
@@ -145,6 +186,32 @@ class NuxtDevServer extends EventEmitter {
         },
         ...this.options.overrides,
       },
+    })
+
+    // Connect Vite HMR
+    if (!process.env.NUXI_DISABLE_VITE_HMR) {
+      this._currentNuxt.hooks.hook(
+        'vite:extendConfig',
+        (config, { isClient }) => {
+          if (isClient && config.server) {
+            config.server.hmr = {
+              ...(config.server.hmr as Exclude<
+                typeof config.server.hmr,
+                boolean
+              >),
+              protocol: undefined,
+              port: undefined,
+              host: undefined,
+              server: this.listener.server,
+            }
+          }
+        },
+      )
+    }
+
+    // Remove websocket handlers on close
+    this._currentNuxt.hooks.hookOnce('close', () => {
+      this.listener.server.removeAllListeners('upgrade')
     })
 
     // Write manifest and also check if we need cache invalidation
@@ -173,6 +240,22 @@ class NuxtDevServer extends EventEmitter {
       await this.load(true)
     })
 
+    if ('upgrade' in this._currentNuxt.server) {
+      this.listener.server.on(
+        'upgrade',
+        async (req: any, socket: any, head: any) => {
+          if (
+            req.url.startsWith(
+              this._currentNuxt?.options.app.buildAssetsDir /* /_nuxt/ */,
+            )
+          ) {
+            return // Skip for Vite HMR
+          }
+          await this._currentNuxt?.server.upgrade(req, socket, head)
+        },
+      )
+    }
+
     await this._currentNuxt.hooks.callHook(
       'listen',
       this.listener.server,
@@ -184,12 +267,18 @@ class NuxtDevServer extends EventEmitter {
     const addr = this.listener.address
     this._currentNuxt.options.devServer.host = addr.address
     this._currentNuxt.options.devServer.port = addr.port
-    this._currentNuxt.options.devServer.url = `http://${
-      addr.address.includes(':') ? `[${addr.address}]` : addr.address
-    }:${addr.port}/`
-    this._currentNuxt.options.devServer.https = this.options.https as
-      | boolean
-      | { key: string; cert: string }
+    this._currentNuxt.options.devServer.url = _getAddressURL(
+      addr,
+      !!this.listener.https,
+    )
+    this._currentNuxt.options.devServer.https = this.options.devContext.proxy
+      ?.https as boolean | { key: string; cert: string }
+
+    if (this.listener.https && !process.env.NODE_TLS_REJECT_UNAUTHORIZED) {
+      consola.warn(
+        'You might need `NODE_TLS_REJECT_UNAUTHORIZED=0` environment vairable to make https work.',
+      )
+    }
 
     await Promise.all([
       kit.writeTypes(this._currentNuxt).catch(console.error),
@@ -210,10 +299,13 @@ class NuxtDevServer extends EventEmitter {
   }
 
   async _watchConfig() {
-    const configWatcher = chokidar.watch([this.options.cwd], {
-      ignoreInitial: true,
-      depth: 0,
-    })
+    const configWatcher = chokidar.watch(
+      [this.options.cwd, join(this.options.cwd, '.config')],
+      {
+        ignoreInitial: true,
+        depth: 0,
+      },
+    )
     configWatcher.on('all', (_event, _file) => {
       const file = relative(this.options.cwd, _file)
       if (file === (this.options.dotenv || '.env')) {
@@ -224,4 +316,14 @@ class NuxtDevServer extends EventEmitter {
       }
     })
   }
+}
+
+function _getAddressURL(addr: AddressInfo, https: boolean) {
+  const proto = https ? 'https' : 'http'
+  let host = addr.address.includes(':') ? `[${addr.address}]` : addr.address
+  if (host === '[::]') {
+    host = 'localhost' // Fix issues with Docker networking
+  }
+  const port = addr.port || 3000
+  return `${proto}://${host}:${port}/`
 }