+++ src/utils/kit.ts
@@ -7,7 +7,7 @@ export const loadKit = async (
   rootDir: string,
 ): Promise<typeof import('@nuxt/kit')> => {
   try {
-    // Without PNP (or if users have a local install of kit, we bypass resolving from nuxt)
+    // Without PNP (or if users have a local install of kit, we bypass resolving from Nuxt)
     const localKit = await tryResolveModule('@nuxt/kit', rootDir)
     // Otherwise, we resolve Nuxt _first_ as it is Nuxt's kit dependency that will be used
     const rootURL = localKit ? rootDir : (await tryResolveNuxt()) || rootDir
@@ -31,7 +31,7 @@ export const loadKit = async (
 }
 
 async function tryResolveNuxt() {
-  for (const pkg of ['nuxt3', 'nuxt', 'nuxt-edge']) {
+  for (const pkg of ['nuxt-nightly', 'nuxt3', 'nuxt', 'nuxt-edge']) {
     const path = await tryResolveModule(pkg)
     if (path) {
       return path